+++ ci/syntax/d/ifs.d
@@ -1,0 +1,12 @@
+void main() {
+	if (x) {
+	}
+	if (x)
+	{}
+
+	while (x) {}
+	for (x; y; z) {}
+	foreach (x, y; z) {}
+	switch (x) {}
+	final switch (x) {}
+}